requirements.txt
Package                  Version
------------------------ ---------
alembic                  1.10.4
anyio                    3.6.2
appnope                  0.1.3
asttokens                2.2.1
attrs                    22.2.0
backcall                 0.2.0
black                    23.3.0
boto3                    1.26.102
botocore                 1.29.102
CacheControl             0.12.11
cachetools               5.3.0
certifi                  2022.12.7
cffi                     1.15.1
charset-normalizer       3.1.0
click                    8.1.3
comm                     0.1.3
cryptography             40.0.2
debugpy                  1.6.7
decorator                5.1.1
exceptiongroup           1.1.1
executing                1.2.0
fastapi                  0.95.0
firebase-admin           6.1.0
google-api-core          2.11.0
google-api-python-client 2.85.0
google-auth              2.17.3
google-auth-httplib2     0.1.0
google-cloud-core        2.3.2
google-cloud-firestore   2.11.0
google-cloud-storage     2.8.0
google-crc32c            1.5.0
google-resumable-media   2.4.1
googleapis-common-protos 1.59.0
greenlet                 2.0.2
grpcio                   1.53.0
grpcio-status            1.53.0
h11                      0.14.0
httpcore                 0.16.3
httplib2                 0.22.0
httpx                    0.23.3
idna                     3.4
importlib-metadata       6.3.0
iniconfig                2.0.0
ipykernel                6.22.0
ipython                  8.12.0
jedi                     0.18.2
jmespath                 1.0.1
jupyter-client           8.1.0
jupyter-core             5.3.0
Mako                     1.2.4
MarkupSafe               2.1.2
matplotlib-inline        0.1.6
msgpack                  1.0.5
mypy-extensions          1.0.0
nest-asyncio             1.5.6
numpy                    1.24.2
packaging                23.0
pandas                   1.5.3
parso                    0.8.3
pathspec                 0.11.1
pexpect                  4.8.0
pickleshare              0.7.5
Pillow                   9.4.0
pip                      21.1.3
platformdirs             3.2.0
pluggy                   1.0.0
prompt-toolkit           3.0.38
proto-plus               1.22.2
protobuf                 4.22.3
psutil                   5.9.4
psycopg2-binary          2.9.6
ptyprocess               0.7.0
pure-eval                0.2.2
pyasn1                   0.4.8
pyasn1-modules           0.2.8
pycparser                2.21
pydantic                 1.10.7
Pygments                 2.15.0
PyJWT                    2.6.0
pyparsing                3.0.9
pytest                   7.2.2
python-dateutil          2.8.2
python-multipart         0.0.6
pytz                     2023.3
PyYAML                   6.0
pyzmq                    25.0.2
requests                 2.28.2
rfc3986                  1.5.0
rsa                      4.9
s3transfer               0.6.0
setuptools               56.0.0
six                      1.16.0
sniffio                  1.3.0
SQLAlchemy               2.0.11
stack-data               0.6.2
starlette                0.26.1
tomli                    2.0.1
tornado                  6.2
traitlets                5.9.0
typing-extensions        4.5.0
uritemplate              4.1.1
urllib3                  1.26.15
uvicorn                  0.21.1
wcwidth                  0.2.6
zipp                     3.15.0
